pupple_backend/requirements.txt
absl-py==1.0.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
black==21.12b0
bleach==4.1.0
bs4==0.0.1
cachetools==5.1.0
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.3
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==3.2
django-photologue==3.15.1
django-sortedm2m==3.1.1
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
filelock==3.4.0
flatbuffers==1.12
fonttools==4.29.1
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
imageio==2.16.1
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
krwordrank==1.0.3
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.8
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pafy==0.5.5
pandoc==2.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2021.11.23
platformdirs==2.4.0
plumbum==1.7.2
ply==3.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.1.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
termcolor==1.1.0
terminado==0.13.2
testpath==0.6.0
Theano==1.0.5
threadpoolctl==3.1.0
tifffile==2022.3.25
tomli==1.2.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.1.1
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.5.2
wrapt==1.14.1
youtube-dl==2021.12.17
ExifRead==3.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
Pillow==9.1.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
